added projects endpoint file
